requirements.txt
Pillow>=11.0.0
colorama>=0.4.6
tqdm>=4.65.0
